% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{AdsenseReportsGenerateResponse.headers}
\alias{AdsenseReportsGenerateResponse.headers}
\title{AdsenseReportsGenerateResponse.headers Object}
\usage{
AdsenseReportsGenerateResponse.headers()
}
\value{
AdsenseReportsGenerateResponse.headers object
}
\description{
AdsenseReportsGenerateResponse.headers Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AdsenseReportsGenerateResponse functions: \code{\link{AdsenseReportsGenerateResponse}}
}

