% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{AdsenseReportsGenerateResponse}
\alias{AdsenseReportsGenerateResponse}
\title{AdsenseReportsGenerateResponse Object}
\usage{
AdsenseReportsGenerateResponse(AdsenseReportsGenerateResponse.headers = NULL,
  averages = NULL, endDate = NULL, headers = NULL, rows = NULL,
  startDate = NULL, totalMatchedRows = NULL, totals = NULL,
  warnings = NULL)
}
\arguments{
\item{AdsenseReportsGenerateResponse.headers}{The \link{AdsenseReportsGenerateResponse.headers} object or list of objects}

\item{averages}{The averages of the report}

\item{endDate}{The requested end date in yyyy-mm-dd format}

\item{headers}{The header information of the columns requested in the report}

\item{rows}{The output rows of the report}

\item{startDate}{The requested start date in yyyy-mm-dd format}

\item{totalMatchedRows}{The total number of rows matched by the report request}

\item{totals}{The totals of the report}

\item{warnings}{Any warnings associated with generation of the report}
}
\value{
AdsenseReportsGenerateResponse object
}
\description{
AdsenseReportsGenerateResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AdsenseReportsGenerateResponse functions: \code{\link{AdsenseReportsGenerateResponse.headers}}
}

