% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{AdStyle}
\alias{AdStyle}
\title{AdStyle Object}
\usage{
AdStyle(AdStyle.colors = NULL, AdStyle.font = NULL, colors = NULL,
  corners = NULL, font = NULL)
}
\arguments{
\item{AdStyle.colors}{The \link{AdStyle.colors} object or list of objects}

\item{AdStyle.font}{The \link{AdStyle.font} object or list of objects}

\item{colors}{The colors which are included in the style}

\item{corners}{The style of the corners in the ad}

\item{font}{The font which is included in the style}
}
\value{
AdStyle object
}
\description{
AdStyle Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AdStyle functions: \code{\link{AdStyle.colors}},
  \code{\link{AdStyle.font}}
}

