% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{AdUnit.feedAdsSettings}
\alias{AdUnit.feedAdsSettings}
\title{AdUnit.feedAdsSettings Object}
\usage{
AdUnit.feedAdsSettings(adPosition = NULL, frequency = NULL,
  minimumWordCount = NULL, type = NULL)
}
\arguments{
\item{adPosition}{The position of the ads relative to the feed entries}

\item{frequency}{The frequency at which ads should appear in the feed (i}

\item{minimumWordCount}{The minimum length an entry should be in order to have attached ads}

\item{type}{The type of ads which should appear}
}
\value{
AdUnit.feedAdsSettings object
}
\description{
AdUnit.feedAdsSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Settings specific to feed ads (AFF) - deprecated.
}
\seealso{
Other AdUnit functions: \code{\link{AdUnit.contentAdsSettings.backupOption}},
  \code{\link{AdUnit.contentAdsSettings}},
  \code{\link{AdUnit.mobileContentAdsSettings}},
  \code{\link{AdUnit}}
}

