% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{AdUnit.mobileContentAdsSettings}
\alias{AdUnit.mobileContentAdsSettings}
\title{AdUnit.mobileContentAdsSettings Object}
\usage{
AdUnit.mobileContentAdsSettings(markupLanguage = NULL,
  scriptingLanguage = NULL, size = NULL, type = NULL)
}
\arguments{
\item{markupLanguage}{The markup language to use for this ad unit}

\item{scriptingLanguage}{The scripting language to use for this ad unit}

\item{size}{Size of this ad unit}

\item{type}{Type of this ad unit}
}
\value{
AdUnit.mobileContentAdsSettings object
}
\description{
AdUnit.mobileContentAdsSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Settings specific to WAP mobile content ads (AFMC) - deprecated.
}
\seealso{
Other AdUnit functions: \code{\link{AdUnit.contentAdsSettings.backupOption}},
  \code{\link{AdUnit.contentAdsSettings}},
  \code{\link{AdUnit.feedAdsSettings}},
  \code{\link{AdUnit}}
}
