% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{AdUnit}
\alias{AdUnit}
\title{AdUnit Object}
\usage{
AdUnit(AdUnit.contentAdsSettings = NULL,
  AdUnit.contentAdsSettings.backupOption = NULL,
  AdUnit.feedAdsSettings = NULL, AdUnit.mobileContentAdsSettings = NULL,
  code = NULL, contentAdsSettings = NULL, customStyle = NULL,
  feedAdsSettings = NULL, id = NULL, mobileContentAdsSettings = NULL,
  name = NULL, savedStyleId = NULL, status = NULL)
}
\arguments{
\item{AdUnit.contentAdsSettings}{The \link{AdUnit.contentAdsSettings} object or list of objects}

\item{AdUnit.contentAdsSettings.backupOption}{The \link{AdUnit.contentAdsSettings.backupOption} object or list of objects}

\item{AdUnit.feedAdsSettings}{The \link{AdUnit.feedAdsSettings} object or list of objects}

\item{AdUnit.mobileContentAdsSettings}{The \link{AdUnit.mobileContentAdsSettings} object or list of objects}

\item{code}{Identity code of this ad unit, not necessarily unique across ad clients}

\item{contentAdsSettings}{Settings specific to content ads (AFC) and highend mobile content ads (AFMC - deprecated)}

\item{customStyle}{Custom style information specific to this ad unit}

\item{feedAdsSettings}{Settings specific to feed ads (AFF) - deprecated}

\item{id}{Unique identifier of this ad unit}

\item{mobileContentAdsSettings}{Settings specific to WAP mobile content ads (AFMC) - deprecated}

\item{name}{Name of this ad unit}

\item{savedStyleId}{ID of the saved ad style which holds this ad unit's style information}

\item{status}{Status of this ad unit}
}
\value{
AdUnit object
}
\description{
AdUnit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AdUnit functions: \code{\link{AdUnit.contentAdsSettings.backupOption}},
  \code{\link{AdUnit.contentAdsSettings}},
  \code{\link{AdUnit.feedAdsSettings}},
  \code{\link{AdUnit.mobileContentAdsSettings}}
}

