% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_functions.R
\name{adunits.getAdCode}
\alias{adunits.getAdCode}
\title{Get ad code for the specified ad unit.}
\usage{
adunits.getAdCode(adClientId, adUnitId)
}
\arguments{
\item{adClientId}{Ad client with contains the ad unit}

\item{adUnitId}{Ad unit to get the code for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adsense
\item https://www.googleapis.com/auth/adsense.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adsense, https://www.googleapis.com/auth/adsense.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/adsense/management/}{Google Documentation}
}
