% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_functions.R
\name{alerts.delete}
\alias{alerts.delete}
\title{Dismiss (delete) the specified alert from the publisher's AdSense account.}
\usage{
alerts.delete(alertId)
}
\arguments{
\item{alertId}{Alert to delete}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adsense
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adsense)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/adsense/management/}{Google Documentation}
}
