% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{CustomChannel.targetingInfo}
\alias{CustomChannel.targetingInfo}
\title{CustomChannel.targetingInfo Object}
\usage{
CustomChannel.targetingInfo(adsAppearOn = NULL, description = NULL,
  location = NULL, siteLanguage = NULL)
}
\arguments{
\item{adsAppearOn}{The name used to describe this channel externally}

\item{description}{The external description of the channel}

\item{location}{The locations in which ads appear}

\item{siteLanguage}{The language of the sites ads will be displayed on}
}
\value{
CustomChannel.targetingInfo object
}
\description{
CustomChannel.targetingInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The targeting information of this custom channel, if activated.
}
\seealso{
Other CustomChannel functions: \code{\link{CustomChannel}}
}

