% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_functions.R
\name{metadata.dimensions.list}
\alias{metadata.dimensions.list}
\title{List the metadata for the dimensions available to this AdSense account.}
\usage{
metadata.dimensions.list()
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adsense
\item https://www.googleapis.com/auth/adsense.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adsense, https://www.googleapis.com/auth/adsense.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details. 

#' @importFrom googleAuthR gar_api_generator
}
\seealso{
\href{https://developers.google.com/adsense/management/}{Google Documentation}
}

