% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{Metadata}
\alias{Metadata}
\title{Metadata Object}
\usage{
Metadata(items = NULL)
}
\arguments{
\item{items}{No description}
}
\value{
Metadata object
}
\description{
Metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
