% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{Payment}
\alias{Payment}
\title{Payment Object}
\usage{
Payment(id = NULL, paymentAmount = NULL, paymentAmountCurrencyCode = NULL,
  paymentDate = NULL)
}
\arguments{
\item{id}{Unique identifier of this Payment}

\item{paymentAmount}{The amount to be paid}

\item{paymentAmountCurrencyCode}{The currency code for the amount to be paid}

\item{paymentDate}{The date this payment was/will be credited to the user, or none if the payment threshold has not been met}
}
\value{
Payment object
}
\description{
Payment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

