% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{Payments}
\alias{Payments}
\title{Payments Object}
\usage{
Payments(items = NULL)
}
\arguments{
\item{items}{The list of Payments for the account}
}
\value{
Payments object
}
\description{
Payments Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

