% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{SavedAdStyle}
\alias{SavedAdStyle}
\title{SavedAdStyle Object}
\usage{
SavedAdStyle(adStyle = NULL, id = NULL, name = NULL)
}
\arguments{
\item{adStyle}{The AdStyle itself}

\item{id}{Unique identifier of this saved ad style}

\item{name}{The user selected name of this SavedAdStyle}
}
\value{
SavedAdStyle object
}
\description{
SavedAdStyle Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
