% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{SavedAdStyles}
\alias{SavedAdStyles}
\title{SavedAdStyles Object}
\usage{
SavedAdStyles(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of this response for caching purposes}

\item{items}{The saved ad styles returned in this list response}

\item{nextPageToken}{Continuation token used to page through ad units}
}
\value{
SavedAdStyles object
}
\description{
SavedAdStyles Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

