% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{SavedReport}
\alias{SavedReport}
\title{SavedReport Object}
\usage{
SavedReport(id = NULL, name = NULL)
}
\arguments{
\item{id}{Unique identifier of this saved report}

\item{name}{This saved report's name}
}
\value{
SavedReport object
}
\description{
SavedReport Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
