% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{UrlChannel}
\alias{UrlChannel}
\title{UrlChannel Object}
\usage{
UrlChannel(id = NULL, urlPattern = NULL)
}
\arguments{
\item{id}{Unique identifier of this URL channel}

\item{urlPattern}{URL Pattern of this URL channel}
}
\value{
UrlChannel object
}
\description{
UrlChannel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
